requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
black==25.9.0
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
flower==2.0.1
humanize==4.14.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mypy_extensions==1.1.0
openpyxl==3.1.5
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
rabbitmq==0.2.0
referencing==0.37.0
rpds-py==0.29.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
xlrd==2.0.2
